SUSR-102 - hacking Jira user management to not try to add or remove from jira-users group per Wells Fargo issue.
git-svn-id: https://studio.plugins.atlassian.com/svn/SUSR/trunk@78216 2c54a935-e501-0410-bc05-97a93f6bca70

+++ src/main/java/csum/confluence/permissionmgmt/service/impl/JiraSoapUserManagementService.java
@@ -199,12 +199,12 @@ public class JiraSoapUserManagementService extends BaseUserManagementService {
                 }
 
                 if (remoteGroup != null) {
-                    if (!isMemberOf(remoteUser.getName(), groupName)) {
+                    if (!isMemberOf(remoteUser.getName(), groupName) || "jira-users".equals(groupName)) {
                         log.debug("Adding " + remoteUser.getName() + " to Jira group " + groupName);
                         jiraSoapService.addUserToGroup(token, remoteGroup, remoteUser);
                     }
                     else {
-                        log.debug("User " + remoteUser.getName() + " was already a member of Jira group " + groupName);
+                        log.debug("User " + remoteUser.getName() + " was already a member of Jira group " + groupName + ", or group name was 'jira-users'.");
                     }
                 }
             }
@@ -299,12 +299,12 @@ public class JiraSoapUserManagementService extends BaseUserManagementService {
                 }
 
                 if (remoteGroup != null) {
-                    if (isMemberOf(remoteUser.getName(), groupName)) {
+                    if (isMemberOf(remoteUser.getName(), groupName) && !"jira-users".equals(groupName)) {
                         log.debug("Removing " + remoteUser.getName() + " from Jira group " + groupName);
                         jiraSoapService.removeUserFromGroup(token, remoteGroup, remoteUser);
                     }
                     else {
-                        log.debug("User " + remoteUser.getName() + " was not a member of Jira group " + groupName + " so did not have to remove.");
+                        log.debug("User " + remoteUser.getName() + " was not a member of Jira group " + groupName + " so did not have to remove, or group name was 'jira-users'.");
                     }
                 }
             }
